requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiolimiter==1.2.1
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.63.0
anyio==4.10.0
asn1crypto==1.5.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
backports.asyncio.runner==1.2.0
beartype==0.21.0
blinker==1.9.0
blis==1.3.0
blobfile==3.0.0
boto3==1.40.8
botocore==1.40.8
braintrust==0.2.2
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
chess==1.11.2
chevron==0.14.0
click==8.1.8
cloudpathlib==0.21.1
cloudpickle==3.1.1
colorama==0.4.6
confection==0.1.5
contourpy==1.3.0
cryptography==45.0.6
cycler==0.12.1
cymem==2.0.11
dacite==1.9.2
datasets==4.0.0
deepeval==3.3.9
dill==0.3.8
distro==1.9.0
docker==7.1.0
evals==3.0.1.post1
evaluate==0.4.5
exceptiongroup==1.3.0
execnet==2.1.1
Farama-Notifications==0.0.4
filelock==3.18.0
fire==0.7.0
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.3.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.29.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.71.2
gymnasium==1.1.1
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.7
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jiwer==4.0.0
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
keras==3.10.0
kiwisolver==1.4.7
langchain==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.30
langchain-text-splitters==0.3.9
langcodes==3.5.0
langdetect==1.0.9
langsmith==0.4.14
language_data==1.3.0
libclang==18.1.1
lxml==6.0.0
lz4==4.4.4
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
ml_dtypes==0.5.3
mock==5.2.0
monotonic==1.6
multidict==6.6.4
multiprocess==0.70.16
murmurhash==1.0.13
mypy==1.17.1
mypy_extensions==1.1.0
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numexpr==2.10.2
numpy==2.0.2
ollama==0.5.3
openai==1.99.9
openevals==0.1.0
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
# opentracing==2.4.0
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.2
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
patsy==1.0.1
pillow==11.3.0
platformdirs==4.3.8
playwright==1.54.0
pluggy==1.6.0
portalocker==3.2.0
posthog==3.25.0
preshed==3.0.10
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
pyfiglet==1.0.3
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-repeat==0.9.4
pytest-rerunfailures==12.0
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
rsa==4.9.1
s3transfer==0.13.1
sacrebleu==2.5.1
scipy==1.13.1
seaborn==0.13.2
sentry-sdk==2.34.1
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
snowflake-connector-python==3.16.0
sortedcontainers==2.4.0
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy_universal_sentence_encoder==0.4.6
SQLAlchemy==2.0.43
srsly==2.5.1
sseclient-py==1.8.0
statsmodels==0.14.5
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
text-unidecode==1.3
tf_keras==2.19.0
thinc==8.3.6
tiktoken==0.11.0
tomli==2.2.1
tomlkit==0.13.3
tqdm==4.67.1
typer==0.16.0
types-PyYAML==6.0.12.20250809
types-requests==2.31.0.6
types-tqdm==4.67.0.20250809
types-urllib3==1.26.25.14
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==1.26.20
wasabi==1.1.3
weasel==0.4.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
